+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.5.5
+
+* Fixed a problem with Connect-Graph not passing dynamic parameters (like Certificate) to Connect-MgGraph
+* Fixed a bug in Get-GraphGroup which prevented owners of a group being returned (small typo)
+* Added control of user properties fetched for owners and members of groups in Get-GraphGroup
+
+
 ## 1.5.4
 
 * Added Support for -UseDeviceAuthentication